# Conflicts:
#   config.yaml.sample
#   go.mod
#   go.sum
#   internal/config/main.go
#   main.go

+++ internal/modules/todo/service.go
@@ -1,6 +1,9 @@
 package todo
 
-import "fmt"
+import (
+	"fmt"
+	"golang-api-starter/internal/helper"
+)
 
 type Service struct {
 	repo *Repository
@@ -10,9 +13,19 @@ func NewService(r *Repository) *Service {
 	return &Service{r}
 }
 
-func (s *Service) Get(queries map[string]interface{}) []*Todo {
+func (s *Service) Get(queries map[string]interface{}) ([]*Todo, *helper.Pagination) {
 	fmt.Printf("todo service\n")
 	return s.repo.Get(queries)
+}
+
+func (s *Service) GetById(queries map[string]interface{}) ([]*Todo, error) {
+	fmt.Printf("todo service\n")
+
+	records,_ := s.repo.Get(queries)
+	if len(records) == 0 {
+		return nil, fmt.Errorf("%s with id: %s not found", tableName, queries["id"])
+	}
+	return records, nil
 }
 
 func (s *Service) Create(todos []*Todo) []*Todo {
@@ -26,5 +39,13 @@ func (s *Service) Update(todos []*Todo) []*Todo {
 }
 
 func (s *Service) Delete(ids *[]int64) ([]*Todo, error) {
+	idsString, _ := helper.ConvertNumberSliceToString(*ids)
+	records,_ := s.repo.Get(map[string]interface{}{
+		"id": idsString,
+	})
+	if len(records) == 0 {
+		return nil, fmt.Errorf("failed to delete, %s with id: %+v not found", tableName, ids)
+	}
+
 	return s.repo.Delete(ids)
 }
